Request episode watch times from WTC server
The watch time of an episode can be used to determine the watched
status of the episode. It will be possible to filter watched episodes
using that watched status.

The Watch Time Collector (WTC) server allows for requests to be sent in
batches using a comma delimited list of episode IDs. Since the Recently
Added page loads many episodes at once the request is batched rather
than sending individual requests for each episode.