Disable Enhanced RT options on settings page
Rooster Teeth has added an auto play setting which makes the Enhanced RT
binge mode setting unnecessary. There is no need for Enhanced RT options
on the settings page at this time since binge mode was the only setting
available. For now, this is only disabled in case there is a use for the
settings page in the future.