Fix misuse of episode div property for data storage
The id property of each episode's div were being used to store the
channel id that the episode is associated with. This was an incorrect
use of the id property. The id property must be unique but many of the
episode div's were set to the same channel id. Rooster Teeth might use
the id property in the future so it is best to not use it at all.
Episode data is now stored in the div's data attributes.