Disable watched status label and watched filter
The WTC server response time is slow so it takes awhile to determine
watched status. Some times the WTC server returns 502 bad gateway and
therefore watched status cannot be determined. Research needs to be done
to see if this can be improved. Look into fetch connection instead of
xhr. For now it is best to disable this feature.