Add channel color accents to Recently Added page
Channel color accents have been added to each episode shown on the
Recently Added page in order to clearly indicate the channel that made
the content. Color code and logo URL for each channel has been hard
coded into the script because those items are not yet exposed in the API.
Channel colors and logos were sourced from the individual channel pages
on the RT website.