Enable watched status and add resume status indicators
Now that the WTC server is more reliable the watched status labels can be re-enabled. The watched status threshold is now 35 seconds. Episodes will be marked as resumable if they have a watch time but do not meet the watched threshold. Resumable episodes will have a progress percentage bar below the thumbnail.